feat: Refactor countdown app with sound toggle, theme toggle, and quote management
